Move to webpack
A few months later after the systemjs version was made, webpack 2+ seems to have matured a lot more and I believe is now better and simpler than the older systemjs method.

This is largely in part to the webpack dev middleware, which is only available in aspnet core and doesn't work in older aspnet 4 projects. For these, the systemjs branch still available and the yeoman 1.x generator will still generate the systemjs version, but going forward this appears to be the future.

I've also added @angular/material and @covalent/core, which will be toggles in the yeoman generator